Moved set_global_thread_count into its own "threads" module
